feat: adjust code
