+++ static/index.js
@@ -1,158 +1,49 @@
-// 建立list_bar
-function createMrtList(data){
-	let container = document.getElementById("mrts_container");
+import * as divBuilder from './div-builder.js' 
 
-	data.forEach((mrt_name) => {
-		let div =document.createElement("button");
-		div.textContent = mrt_name;
-		container.appendChild(div);
-		div.classList.add("mrt_button", "body-med")
-	});
-}
-
-
-// 網頁開啟時，載入list_bar
-fetch(`/api/mrts`, {
-    method: "GET",
-    headers: {
-        "Content-Type": "application/json",
-    },
-})
-.then(respose => respose.json())
-.then(function(responseData) {
-    const mrt_list = responseData.data;
-
-	const filteredMrtList = mrt_list.filter(item => item !== null);
-
-	createMrtList(filteredMrtList)
-
-	mrtButtons = document.querySelectorAll(".mrt_button");
-
-	mrtButtons.forEach(button => {
-		button.addEventListener("click", function () {
-			const buttonText = this.textContent;
-			
-			searchQuery = buttonText;
-			fetchAndDisplayAttractions(searchQuery);
-
-			searchBox = document.getElementById("searchBox");
-   			searchBox.value = "";
-    		searchBox.setAttribute("placeholder", buttonText);
-			
-			
-		})
-	});
-});
-
-
-// list_bar左右捲動
-const leftBtn = document.getElementById("left_btn")
-const rightBtn = document.getElementById("right_btn")
-
-leftBtn.addEventListener("click", scrollLeft);
-rightBtn.addEventListener("click", scrollRight);
-
-function scrollLeft() {
-    mrtsContainer = document.getElementById("mrts_container");
-    const currentScrollLeft = mrtsContainer.scrollLeft;
-    const scrollAmount = mrtsContainer.clientWidth - 20;
-    const newScrollLeft = currentScrollLeft - scrollAmount;
-    mrtsContainer.scroll({top: 0, left: newScrollLeft, behavior: "smooth"});
-}
-
-function scrollRight() {
-    mrtsContainer = document.getElementById("mrts_container");
-    const currentScrollLeft = mrtsContainer.scrollLeft;
-    const scrollAmount = mrtsContainer.clientWidth - 20;
-    const newScrollLeft = currentScrollLeft + scrollAmount;
-    mrtsContainer.scroll({top: 0, left: newScrollLeft, behavior: "smooth"});
-}
-
-
-// 建立查無資料
-function createNoData(){
-	let container = document.getElementById("attractions");
-	attractionNoData = document.createElement("div");
-	attractionNoData.classList.add("attraction_nodata", "content-bold");
-	attractionNoData.textContent = "查無相關景點"
-
-	container.appendChild(attractionNoData);
-}
-
-// 建立attractions
-function createAttraction(data) {
-	let container = document.getElementById("attractions");
-
-	data.forEach((item) => {
-		let attraction_name = item.name
-		let attraction_category = item.category
-		let attraction_mrt = item.mrt
-		let attraction_image = item.images[0]
-
-		// 創建包裝每個景點的<div>
-		let attractionContainer = document.createElement("div");
-		attractionContainer.classList.add("attraction_wrapper"); // 添加額外的class
-
-	
-		// 創建圖片元素
-		let img = document.createElement("img");
-		img.src = attraction_image;
-
-		// 創建景點名稱元素
-		let nameDiv = document.createElement("div");
-		nameDiv.textContent = attraction_name;
-		nameDiv.classList.add("attraction_name", "body-bold")
-
-		// 創建捷運名稱元素
-		let attractionMrtDiv = document.createElement("div");
-		attractionMrtDiv.textContent = attraction_mrt;
-		attractionMrtDiv.classList.add("attraction_mrt", "body-med");
-
-		// 創建特徵元素
-		let categoryDiv = document.createElement("div");
-		categoryDiv.textContent = attraction_category;
-		categoryDiv.classList.add("attraction_category", "body-med");
-
-		// 創建描述容器
-		let describeContainer = document.createElement("div");
-		describeContainer.classList.add("describe");
-
-		// 將捷運名稱和特徵添加到描述容器
-		describeContainer.appendChild(attractionMrtDiv);
-		describeContainer.appendChild(categoryDiv);
-
-		// 創建圖片容器
-		let infoContainer = document.createElement("div");
-		infoContainer.classList.add("img_container");
-
-		// 將圖片和景點名稱添加到圖片容器
-		infoContainer.appendChild(img);
-		infoContainer.appendChild(nameDiv);
-
-		// 將圖片容器和描述容器添加到attractions_wrapper中
-		attractionContainer.appendChild(infoContainer);
-		attractionContainer.appendChild(describeContainer);
-
-		// 將attractions_wrapper添加到attractions中
-		container.appendChild(attractionContainer);
-
-	})
-}
-
-// 返回初始畫面
-titleButton = document.getElementById("title");
-titleButton.addEventListener("click", function(){
-	clearCurrentContent();
-	resetSearchBox();
-	loadInitialData();
-});
 
 let nextPage = null; 
 let isLoading = false;
 let searchQuery = '';
 
 
-// 網頁開啟時，載入初始景點
+
+// api連接-MRT
+function loadMRTList() {
+    return fetch(`/api/mrts`, {
+        method: "GET",
+        headers: {
+            "Content-Type": "application/json",
+        },
+    })
+    .then(respose => respose.json())
+    .then(function(responseData) {
+        const mrt_list = responseData.data;
+        const filteredMrtList = mrt_list.filter(item => item !== null);
+        return filteredMrtList;
+    });
+};
+
+// 監聽器-list_bar
+function mrtButtonHandler() {
+    let mrtButtons = document.querySelectorAll(".mrt_button");
+    mrtButtons.forEach(button => {
+        button.addEventListener("click", function () {
+            const buttonText = this.textContent;
+            
+            let searchQuery = buttonText;
+            clearCurrentContent();
+            fetchAndDisplayAttractions(searchQuery);
+
+            // 更換搜尋欄顯示文字
+            let searchBox = document.getElementById("searchBox");
+                searchBox.value = "";
+            searchBox.setAttribute("placeholder", buttonText);			
+        });
+    });
+};
+
+
+// api連接-attractions
 function loadInitialData() {
 	fetch(`/api/attractions`, {
 		method: "GET",
@@ -162,10 +53,10 @@ function loadInitialData() {
 	})
 	.then(respose => respose.json())
 	.then(function(responseData) {
-		const application_list = responseData.data;
+		let application_list = responseData.data;
 		nextPage = responseData.nextPage;
 		searchQuery = null;
-		createAttraction(application_list)
+		divBuilder.createAttraction(application_list)
 	})
 }
 
@@ -185,17 +76,16 @@ function clearCurrentContent() {
 }
 
 
-// 景點查詢
-searchButton = document.getElementById("searchButton");
+// 景點查詢按鈕
+const searchButton = document.getElementById("searchButton");
 searchButton.addEventListener("click", function(){
 	searchQuery = document.getElementById("searchBox").value;
+    clearCurrentContent();
 	fetchAndDisplayAttractions(searchQuery);
 });
 
 // 關鍵字查詢
 function fetchAndDisplayAttractions(searchQuery) {
-    clearCurrentContent();
-
     fetch(`/api/attractions?keyword=${searchQuery}`, {
         method: "GET",
         headers: {
@@ -206,10 +96,10 @@ function fetchAndDisplayAttractions(searchQuery) {
     .then(function(responseData) {
         const attractionList = responseData.data;
         if (attractionList.length === 0) {
-            createNoData();
+            divBuilder.createNoData();
         } else {
             nextPage = responseData.nextPage;
-            createAttraction(attractionList);
+            divBuilder.createAttraction(attractionList);
         }
     });
 }
@@ -247,7 +137,7 @@ function loadMore(searchQuery){
         .then(function (responseData) {
             const attractionList = responseData.data;
 			
-			createAttraction(attractionList);
+			divBuilder.createAttraction(attractionList);
 
 			nextPage = responseData.nextPage;
 
@@ -262,6 +152,14 @@ function loadMore(searchQuery){
 }
 
 
+// 返回初始畫面
+let titleButton = document.getElementById("title");
+titleButton.addEventListener("click", function(){
+	clearCurrentContent();
+	resetSearchBox();
+	loadInitialData();
+});
+
 // 清空搜尋框
 function resetSearchBox() {
     const searchBox = document.getElementById("searchBox");
@@ -269,5 +167,37 @@ function resetSearchBox() {
     searchBox.setAttribute("placeholder", "輸入景點名稱查詢");
 }
 
+// listBar左右滾動
+function listBarScroll() {
+    let leftBtn = document.getElementById("left_btn")
+    let rightBtn = document.getElementById("right_btn")
 
-loadInitialData()
+    leftBtn.addEventListener("click", () => scroll('left'));
+    rightBtn.addEventListener("click", () => scroll('right'));
+}
+
+function scroll(direction) {
+    const mrtsContainer = document.getElementById("mrts_container");
+    const currentScrollLeft = mrtsContainer.scrollLeft;
+    const scrollAmount = mrtsContainer.clientWidth - 20;
+    let newScrollLeft;
+
+    if (direction === 'left') {
+        newScrollLeft = currentScrollLeft - scrollAmount;
+    }
+    else if (direction === 'right') {
+        newScrollLeft = currentScrollLeft + scrollAmount;
+    }
+    mrtsContainer.scroll({top: 0, left: newScrollLeft, behavior: "smooth"});
+}
+
+
+loadInitialData();
+
+loadMRTList().then(filteredMrtList => {
+    divBuilder.createMrtList(filteredMrtList);
+    mrtButtonHandler();
+});
+
+// list_bar左右捲動
+listBarScroll();
