+++ templates/index.html
@@ -65,6 +65,6 @@
 		<footer class="footer">
 			<div class="container body-bold">COPYRIGHT © 2021 台北一日遊</div>
 		</footer>
-		<script defer type="text/javascript" src="../static/index.js"></script>
+		<script defer type="module" src="../static/index.js"></script>
 	</body>
 </html>
